Changed to VS2019, fixed missing selectToken
