--- AssemblyInfo1.cs
@@ -1,8 +1,0 @@
-using MelonLoader;
-using Movement_Speed_Cap_Plus;
-using System.Reflection;
-
-[assembly: MelonInfo(typeof(MoveSpeedCap), "Move Speed Cap+", "1.0.0", "Farukon")]
-[assembly: MelonGame("poncle", "VampireSurvivors")]
-[assembly: VerifyLoaderVersion(0, 6, 1, true)]
-[assembly: MelonAuthorColor(255, 0, 255, 0)]
