+++ MoveSpeedCap+.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33516.290
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.33529.622
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Movement_Speed_Cap_Plus", "Movement_Speed_Cap_Plus.csproj", "{77E1543F-7FF0-43CB-9AC5-19C6CBD848FB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoveSpeedCap+", "MoveSpeedCap+.csproj", "{78546F5D-6D40-44FF-B56E-9664D4F9888E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{77E1543F-7FF0-43CB-9AC5-19C6CBD848FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{77E1543F-7FF0-43CB-9AC5-19C6CBD848FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{77E1543F-7FF0-43CB-9AC5-19C6CBD848FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{77E1543F-7FF0-43CB-9AC5-19C6CBD848FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{78546F5D-6D40-44FF-B56E-9664D4F9888E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78546F5D-6D40-44FF-B56E-9664D4F9888E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78546F5D-6D40-44FF-B56E-9664D4F9888E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78546F5D-6D40-44FF-B56E-9664D4F9888E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2D76E72D-5C5A-4ADD-81AC-BD42D3FA75AC}
+		SolutionGuid = {7221B1E4-2E3C-4E2C-9400-2640A0488D92}
 	EndGlobalSection
 EndGlobal
