+++ Properties/AssemblyInfo.cs
@@ -1,0 +1,23 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using MelonLoader;
+using Movement_Speed_Cap_Plus;
+
+[assembly: AssemblyTitle(ModInfo.Name)]
+[assembly: AssemblyDescription(ModInfo.Description)]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany(ModInfo.Company)]
+[assembly: AssemblyProduct(ModInfo.Name)]
+[assembly: AssemblyCopyright("Copyright © " + ModInfo.Author + " 2023")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: ComVisible(false)]
+[assembly: Guid("78546f5d-6d40-44ff-b56e-9664d4f9888e")]
+[assembly: AssemblyVersion(ModInfo.Version)]
+[assembly: AssemblyFileVersion(ModInfo.Version)]
+
+[assembly: MelonInfo(typeof(MoveSpeedCap), ModInfo.Name, ModInfo.Version, ModInfo.Author)]
+[assembly: MelonGame("poncle", "VampireSurvivors")]
+[assembly: VerifyLoaderVersion(0, 6, 1, true)]
+[assembly: MelonAuthorColor(255, 0, 255, 0)]
